refactor: service, serviceimpl directories
