+++ 고형인/java-racingcar-6/src/main/java/racingcar/service/InputService.java
@@ -1,6 +1,6 @@
-package racingcar;
+package racingcar.service;
 
 // 입력 인터페이스 클래스
-public interface Input {
+public interface InputService {
     String getInput();
 }
