+++ 고형인/java-racingcar-6/src/main/java/racingcar/serviceimpl/CountInputServiceImpl.java
@@ -1,8 +1,9 @@
-package racingcar;
+package racingcar.serviceimpl;
 
 import camp.nextstep.edu.missionutils.Console;
+import racingcar.service.InputService;
 
-public class CountInput implements Input {
+public class CountInputServiceImpl implements InputService {
     @Override
     public String getInput() {
         System.out.println("시도할 회수는 몇회인가요?");
